requirements.txt
google-auth==2.16.0
google-auth-oauthlib==0.8.0
gspread==5.7.2
jupyter==1.0.0
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.5.2
python-dotenv==0.21.0
selenium==4.7.2
